# openapi.json
{
  "components": {
    "schemas": {
      "AggregationRequest": {
        "description": "聚合分析请求",
        "examples": [
          {
            "group_by": [
              "city"
            ],
            "limit": 5,
            "metrics": [
              "sum"
            ],
            "sort_by": "sum",
            "sort_order": "desc"
          }
        ],
        "properties": {
          "filters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "过滤条件",
            "title": "Filters"
          },
          "group_by": {
            "description": "分组字段",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Group By",
            "type": "array"
          },
          "limit": {
            "anyOf": [
              {
                "maximum": 1000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "限制数量",
            "title": "Limit"
          },
          "metrics": {
            "description": "聚合指标",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Metrics",
            "type": "array"
          },
          "sort_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "排序字段",
            "title": "Sort By"
          },
          "sort_order": {
            "default": "desc",
            "description": "排序顺序",
            "pattern": "^(asc|desc)$",
            "title": "Sort Order",
            "type": "string"
          }
        },
        "required": [
          "group_by",
          "metrics"
        ],
        "title": "AggregationRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "中国城市经济数据 REST API：35 城、16 年指标、集成预测与企业/政府/产业分析。",
    "license": {
      "name": "GPL-3.0-or-later",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    },
    "title": "Urban Pulse",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cities/aggregate": {
      "post": {
        "description": "数据聚合分析(基于真实注册城市数据)\n\n支持按 city_code / year / indicator / province 等字段分组;\nmetrics 支持 count / sum / avg / min / max / median / std。",
        "operationId": "aggregate_data_api_v1_cities_aggregate_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AggregationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "数据聚合分析",
        "tags": [
          "城市"
        ]
      }
    },
...
    "/api/v1/cities/compare": {
      "post": {
        "description": "城市对比分析(基于真实注册城市数据)\n\n对比多个城市在指定指标上的表现;数据空集时返回 503。",
        "operationId": "compare_cities_api_v1_cities_compare_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CitiesCompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "城市对比分析",
        "tags": [
          "城市"
        ]
      }
    },
    "/api/v1/cities/correlation": {
      "post": {
        "description": "指标相关性分析(基于真实注册城市数据)",
        "operationId": "correlation_analysis_api_v1_cities_correlation_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrelationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "相关性分析",
        "tags": [
          "城市"
        ]
      }
    },
    "/api/v1/cities/dashboard": {
      "get": {
        "description": "获取城市仪表盘数据(基于真实注册城市数据)",
        "operationId": "get_city_dashboard_api_v1_cities_dashboard_get",
        "parameters": [
          {
            "description": "城市代码",
            "in": "query",
            "name": "city_code",
            "required": true,
            "schema": {
              "description": "城市代码",
              "minLength": 1,
              "title": "City Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "城市仪表盘",
        "tags": [
          "城市"
        ]
      }
    },
...
    "/api/v1/cities/rankings": {
      "get": {
        "description": "获取城市排名(基于真实注册城市数据)",
        "operationId": "get_city_rankings_api_v1_cities_rankings_get",
        "parameters": [
          {
            "description": "指标代码",
            "in": "query",
            "name": "indicator",
            "required": false,
            "schema": {
              "default": "gdp",
              "description": "指标代码",
              "minLength": 1,
              "title": "Indicator",
              "type": "string"
            }
          },
          {
            "description": "年份",
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2200,
                  "minimum": 1900,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "年份",
              "title": "Year"
            }
          },
          {
            "description": "返回数量",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "description": "返回数量",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "城市排名",
        "tags": [
          "城市"
        ]
      }
    },
    "/api/v1/cities/regional": {
      "post": {
        "description": "区域分析(基于真实注册城市数据)",
        "operationId": "regional_analysis_api_v1_cities_regional_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "区域分析",
        "tags": [
          "城市"
        ]
      }
    },
    "/api/v1/cities/time-series": {
      "post": {
        "description": "时间序列分析(基于真实注册城市数据)",
        "operationId": "time_series_analysis_api_v1_cities_time_series_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeSeriesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "时间序列分析",
        "tags": [
          "城市"
        ]
      }
    },
    "/api/v1/cities/{city_name}": {
      "get": {
        "description": "获取指定城市的详细数据。",
        "operationId": "get_city_detail_api_v1_cities__city_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "city_name",
            "required": true,
            "schema": {
              "title": "City Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get City Detail Api V1 Cities  City Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "获取指定城市详情",
        "tags": [
          "城市"
        ]
      }
    },
...
  },
...
}